move todo.js to router folder
